added auth0 modular example
