+++ web-core-sdk/custom-authentication/auth0-react-modular-example/api/.env.example
@@ -1,0 +1,6 @@
+AUTH_URL = https://test-tor.us.auth0.com/oauth/token
+PORT = 3001
+CLIENT_ID = M7Gutk7odQu3T3VABOq5oEENmZuJFRYt 
+CLIENT_SECRET =  *****
+REDIRECT_URI = http://localhost:3001
+FRONT_ENDPOINT = http://localhost:3000/rwa?token=
