+++ web-core-sdk/custom-authentication/auth0-react-modular-example/api/README.md
@@ -1,0 +1,17 @@
+# Express Server for auth0 RWA example
+
+## Running This Sample Locally
+
+1. Install the dependencies with npm:
+
+```bash
+npm install
+```
+
+2. Rename `.env.example` to `.env` and replace or check the values.
+
+3. Run server
+
+```bash
+node server.js or nodemon server.js
+```
