+++ web-core-sdk/custom-authentication/auth0-react-modular-example/api/package.json
@@ -1,0 +1,17 @@
+{
+  "name": "auth0-express-webapp-sample",
+  "scripts": {
+    "lint": "eslint"
+  },
+  "dependencies": {
+    "debug": "~4.3.4",
+    "dotenv": "^16.0.3",
+    "express": "^4.18.2",
+    "http-errors": "~2.0.0",
+    "request": "^2.88.2"
+  },
+  "devDependencies": {
+    "eslint": "^8.27.0",
+    "nodemon": "^2.0.20"
+  }
+}
