+++ web-core-sdk/custom-authentication/auth0-react-modular-example/app/.env.development
@@ -1,0 +1,7 @@
+REACT_APP_CLIENT_ID="BMuAPXdFXaK94pUgfNluIEBPMTiwWKQz0h8AkCtf4Rzxv4bNLwsTRXSlt5OlB6KSpP_jYFhzloMf2XhUYADB3JE"
+REACT_APP_VERIFIER="web3auth-auth0-demo"
+REACT_APP_DOMAIN="http://localhost:3000"
+REACT_APP_AUTH0_DOMAIN="https://test-tor.us.auth0.com"
+REACT_APP_SPA_CLIENTID="HIlPSdOzq9JrixbWQB4QRcHHsplTU9hX"
+REACT_APP_RWA_CLIENTID="M7Gutk7odQu3T3VABOq5oEENmZuJFRYt"
+REACT_APP_BACKEND_SERVER_API="http://localhost:3001/callback"
