+++ web-core-sdk/custom-authentication/auth0-react-modular-example/app/config-overrides.js
@@ -1,0 +1,45 @@
+const webpack = require("webpack");
+const { BundleAnalyzerPlugin } = require("webpack-bundle-analyzer");
+const path = require("path");
+
+module.exports = function override(config) {
+  const fallback = config.resolve.fallback || {};
+  Object.assign(fallback, {
+    crypto: require.resolve("crypto-browserify"),
+    stream: require.resolve("stream-browserify"),
+    assert: require.resolve("assert"),
+    http: require.resolve("stream-http"),
+    https: require.resolve("https-browserify"),
+    os: require.resolve("os-browserify"),
+    url: require.resolve("url"),
+  });
+  config.resolve.fallback = fallback;
+  config.resolve.alias = {
+    ...config.resolve.alias,
+    "bn.js": path.resolve(__dirname, "node_modules/bn.js"),
+    lodash: path.resolve(__dirname, "node_modules/lodash"),
+  };
+  config.plugins = (config.plugins || []).concat([
+    new webpack.ProvidePlugin({
+      process: "process/browser",
+      Buffer: ["buffer", "Buffer"],
+    }),
+    new webpack.IgnorePlugin({
+      resourceRegExp: /genesisStates\/[a-z]*\.json$/,
+      contextRegExp: /@ethereumjs\/common/,
+    }),
+    new BundleAnalyzerPlugin({
+      analyzerMode: "disabled"
+    }),
+  ]);
+  config.module.rules.push({
+    test: /\.(js|mjs|jsx)$/,
+    enforce: "pre",
+    loader: require.resolve("source-map-loader"),
+    resolve: {
+      fullySpecified: false,
+    },
+  });
+  config.ignoreWarnings = [/Failed to parse source map/];
+  return config;
+};
