+++ web-core-sdk/custom-authentication/auth0-react-modular-example/app/package.json
@@ -1,0 +1,64 @@
+{
+  "name": "auth0-web3auth-example",
+  "version": "0.1.0",
+  "private": true,
+  "dependencies": {
+    "@solana/web3.js": "^1.66.2",
+    "@toruslabs/openlogin": "^2.7.0",
+    "@web3auth/base": "^4.0.0-alpha.0",
+    "@web3auth/core": "^4.0.0-alpha.0",
+    "@web3auth/openlogin-adapter": "^4.0.0-alpha.0",
+    "dotenv": "^16.0.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-loader-spinner": "^5.3.4",
+    "react-router-dom": "^6.4.3",
+    "react-switch": "^7.0.0",
+    "start": "^5.1.0",
+    "web3": "^1.8.0"
+  },
+  "scripts": {
+    "start": "react-app-rewired start",
+    "build": "react-app-rewired build",
+    "lint": "eslint --fix 'src/**/*.ts' 'src/**/*.tsx'",
+    "eject": "react-scripts eject"
+  },
+  "eslintConfig": {
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
+  },
+  "browserslist": {
+    "production": [
+      "chrome >= 67",
+      "edge >= 79",
+      "firefox >= 68",
+      "opera >= 54",
+      "safari >= 14"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  },
+  "devDependencies": {
+    "@types/node": "^18.11.9",
+    "@types/react": "^18.0.25",
+    "@types/react-dom": "^18.0.8",
+    "assert": "^2.0.0",
+    "buffer": "^6.0.3",
+    "crypto-browserify": "^3.12.0",
+    "https-browserify": "^1.0.0",
+    "os-browserify": "^0.3.0",
+    "process": "^0.11.10",
+    "react-app-rewired": "^2.2.1",
+    "react-scripts": "5.0.1",
+    "stream-browserify": "^3.0.0",
+    "stream-http": "^3.2.0",
+    "typescript": "^4.8.4",
+    "url": "^0.11.0",
+    "webpack-bundle-analyzer": "^4.7.0"
+  }
+}
