+++ web-core-sdk/custom-authentication/auth0-react-modular-example/app/public/favicon.ico
