+++ web-core-sdk/custom-authentication/auth0-react-modular-example/app/public/images/github-logo.png
