+++ web-core-sdk/custom-authentication/auth0-react-modular-example/app/public/robots.txt
@@ -1,0 +1,3 @@
+# https://www.robotstxt.org/robotstxt.html
+User-agent: *
+Disallow:
