+++ web-core-sdk/custom-authentication/auth0-react-modular-example/app/src/components/Loader.tsx
@@ -1,0 +1,6 @@
+import { Oval } from "react-loader-spinner";
+function Loader() {
+  return <Oval ariaLabel="loading-indicator" height={100} width={100} strokeWidth={5} color="blue" secondaryColor="grey" />;
+}
+
+export default Loader;
