+++ web-core-sdk/custom-authentication/auth0-react-modular-example/app/src/config/appConfig.ts
@@ -1,0 +1,10 @@
+export const APP_CONFIG = {
+  SPA: {
+    displayName: "SPA",
+  },
+  RWA: {
+    displayName: "RWA",
+  },
+} as const;
+
+export type APP_CONFIG_TYPE = keyof typeof APP_CONFIG;
