+++ web-core-sdk/custom-authentication/auth0-react-modular-example/app/src/config/chainConfig.ts
