+++ web-core-sdk/custom-authentication/auth0-react-modular-example/app/src/config/web3AuthNetwork.ts
@@ -1,0 +1,7 @@
+export const WEB3AUTH_NETWORK = {
+  cyan: {
+    displayName: "Cyan",
+  },
+} as const;
+
+export type WEB3AUTH_NETWORK_TYPE = keyof typeof WEB3AUTH_NETWORK;
