+++ web-core-sdk/custom-authentication/auth0-react-modular-example/app/src/index.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+import { createRoot } from 'react-dom/client'
+import "./index.css";
+import App from "./App";
+
+createRoot(
+  document.getElementById('root') as HTMLElement)
+  .render(
+  <React.StrictMode>
+    <App />
+  </React.StrictMode>
+)
