+++ web-core-sdk/custom-authentication/auth0-react-modular-example/app/src/react-app-env.d.ts
@@ -1,0 +1,1 @@
+/// <reference types="react-scripts" />
