+++ web-core-sdk/custom-authentication/auth0-react-modular-example/app/src/services/walletProvider.ts
