--- web-core-sdk/custom-authentication/firebase-next-core-example/.eslintrc.json
@@ -1,12 +1,0 @@
-{
-  "root": true,
-  "extends": ["next/core-web-vitals", "eslint:recommended", "plugin:react/recommended", "plugin:react-hooks/recommended", "plugin:react/jsx-runtime"],
-  "parserOptions": {
-    "ecmaFeatures": {
-      "jsx": true
-    }
-  },
-  "rules": {
-    "react/require-default-props": 0
-  }
-}
