--- web-core-sdk/custom-authentication/firebase-next-core-example/components/Loader.tsx
@@ -1,7 +1,0 @@
-import { FunctionComponent } from "react";
-
-const Loader: FunctionComponent = () => {
-  return <h1>Loading....</h1>;
-};
-
-export default Loader;
