--- web-core-sdk/custom-authentication/firebase-next-core-example/config/web3AuthNetwork.ts
@@ -1,7 +1,0 @@
-export const WEB3AUTH_NETWORK = {
-  testnet: {
-    displayName: "Testnet",
-  }
-} as const;
-
-export type WEB3AUTH_NETWORK_TYPE = keyof typeof WEB3AUTH_NETWORK;
