--- web-core-sdk/custom-authentication/firebase-next-core-example/next.config.js
@@ -1,9 +1,0 @@
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-  reactStrictMode: true,
-  experimental: {
-    esmExternals: "loose",
-  },
-};
-
-module.exports = nextConfig;
