--- web-core-sdk/custom-authentication/firebase-next-core-example/services/web3auth.tsx
@@ -1,237 +1,0 @@
-import { ADAPTER_EVENTS, SafeEventEmitterProvider } from "@web3auth/base";
-import { Web3AuthCore } from "@web3auth/core";
-import { createContext, FunctionComponent, ReactNode, useCallback, useContext, useEffect, useState } from "react";
-import { CHAIN_CONFIG, CHAIN_CONFIG_TYPE } from "../config/chainConfig";
-import { WEB3AUTH_NETWORK_TYPE } from "../config/web3AuthNetwork";
-import { getWalletProvider, IWalletProvider } from "./walletProvider";
-import { OpenloginAdapter } from "@web3auth/openlogin-adapter";
-
-export interface IWeb3AuthContext {
-  web3Auth: Web3AuthCore | null;
-  provider: IWalletProvider | null;
-  isLoading: boolean;
-  chain: string;
-  user: unknown;
-  login: (loginProvider: string, jwtToken: string) => Promise<void>;
-  logout: () => Promise<void>;
-  getUserInfo: () => Promise<any>;
-  signMessage: () => Promise<any>;
-  getAccounts: () => Promise<any>;
-  getBalance: () => Promise<any>;
-  signTransaction: () => Promise<void>;
-  signAndSendTransaction: () => Promise<void>;
-}
-
-export const Web3AuthContext = createContext<IWeb3AuthContext>({
-  web3Auth: null,
-  provider: null,
-  isLoading: false,
-  chain: "",
-  user: null,
-  login: async () => {},
-  logout: async () => {},
-  getUserInfo: async () => {},
-  signMessage: async () => {},
-  getAccounts: async () => {},
-  getBalance: async () => {},
-  signTransaction: async () => {},
-  signAndSendTransaction: async () => {},
-});
-
-export function useWeb3Auth(): IWeb3AuthContext {
-  return useContext(Web3AuthContext);
-}
-
-interface IWeb3AuthProps {
-  children?: ReactNode;
-  web3AuthNetwork: WEB3AUTH_NETWORK_TYPE;
-  chain: CHAIN_CONFIG_TYPE;
-}
-
-export const Web3AuthProvider: FunctionComponent<IWeb3AuthProps> = ({ children, web3AuthNetwork, chain }: IWeb3AuthProps) => {
-  const [web3Auth, setWeb3Auth] = useState<Web3AuthCore | null>(null);
-  const [provider, setProvider] = useState<IWalletProvider | null>(null);
-  const [user, setUser] = useState<unknown | null>(null);
-  const [isLoading, setIsLoading] = useState(false);
-
-  const setWalletProvider = useCallback(
-    (web3authProvider: SafeEventEmitterProvider) => {
-      const walletProvider = getWalletProvider(chain, web3authProvider, uiConsole);
-      setProvider(walletProvider);
-    },
-    [chain]
-  );
-
-  useEffect(() => {
-    const subscribeAuthEvents = (web3auth: Web3AuthCore) => {
-      // Can subscribe to all ADAPTER_EVENTS and LOGIN_MODAL_EVENTS
-      web3auth.on(ADAPTER_EVENTS.CONNECTED, (data: unknown) => {
-        console.log("Yeah!, you are successfully logged in", data);
-        setUser(data);
-        setWalletProvider(web3auth.provider!);
-      });
-
-      web3auth.on(ADAPTER_EVENTS.CONNECTING, () => {
-        console.log("connecting");
-      });
-
-      web3auth.on(ADAPTER_EVENTS.DISCONNECTED, () => {
-        console.log("disconnected");
-        setUser(null);
-      });
-
-      web3auth.on(ADAPTER_EVENTS.ERRORED, (error) => {
-        console.error("some error or user has cancelled login request", error);
-      });
-    };
-
-    const currentChainConfig = CHAIN_CONFIG[chain];
-
-    async function init() {
-      try {
-        const clientId = "BKPxkCtfC9gZ5dj-eg-W6yb5Xfr3XkxHuGZl2o2Bn8gKQ7UYike9Dh6c-_LaXlUN77x0cBoPwcSx-IVm0llVsLA";
-        setIsLoading(true);
-        const web3AuthInstance = new Web3AuthCore({
-          chainConfig: currentChainConfig,
-          clientId,
-        });
-
-        const adapter = new OpenloginAdapter({
-          adapterSettings: {
-            network: web3AuthNetwork,
-            clientId,
-            uxMode: "redirect",
-            loginConfig: {
-              jwt: {
-                name: "Custom Firebase Login",
-                verifier: process.env.REACT_APP_VERIFIER || "web3auth-firebase-demo",
-                typeOfLogin: "jwt",
-                clientId,
-              },
-            },
-          },
-        });
-        web3AuthInstance.configureAdapter(adapter);
-        subscribeAuthEvents(web3AuthInstance);
-        setWeb3Auth(web3AuthInstance);
-        await web3AuthInstance.init();
-      } catch (error) {
-        console.error(error);
-      } finally {
-        setIsLoading(false);
-      }
-    }
-    init();
-  }, [chain, web3AuthNetwork, setWalletProvider]);
-
-  const login = async (loginProvider: string, jwtToken: string) => {
-    setIsLoading(true);
-    try {
-      if (!web3Auth) {
-        console.log("web3auth not initialized yet");
-        uiConsole("web3auth not initialized yet");
-        return;
-      }
-      await web3Auth.connectTo("openlogin", {
-        relogin: true,
-        loginProvider,
-        extraLoginOptions: {
-          id_token: jwtToken,
-          domain: process.env.REACT_APP_DOMAIN || "http://localhost:3000",
-          verifierIdField: "sub",
-        },
-      });
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  const logout = async () => {
-    if (!web3Auth) {
-      console.log("web3auth not initialized yet");
-      uiConsole("web3auth not initialized yet");
-      return;
-    }
-    await web3Auth.logout();
-    setProvider(null);
-  };
-
-  const getUserInfo = async () => {
-    if (!web3Auth) {
-      console.log("web3auth not initialized yet");
-      uiConsole("web3auth not initialized yet");
-      return;
-    }
-    const user = await web3Auth.getUserInfo();
-    uiConsole(user);
-  };
-
-  const getAccounts = async () => {
-    if (!provider) {
-      console.log("provider not initialized yet");
-      uiConsole("provider not initialized yet");
-      return;
-    }
-    await provider.getAccounts();
-  };
-
-  const getBalance = async () => {
-    if (!provider) {
-      console.log("provider not initialized yet");
-      uiConsole("provider not initialized yet");
-      return;
-    }
-    await provider.getBalance();
-  };
-
-  const signMessage = async () => {
-    if (!provider) {
-      console.log("provider not initialized yet");
-      uiConsole("provider not initialized yet");
-      return;
-    }
-    await provider.signMessage();
-  };
-
-  const signTransaction = async () => {
-    if (!provider) {
-      console.log("provider not initialized yet");
-      uiConsole("provider not initialized yet");
-      return;
-    }
-    await provider.signTransaction();
-  };
-
-  const signAndSendTransaction = async () => {
-    if (!provider) {
-      console.log("provider not initialized yet");
-      uiConsole("provider not initialized yet");
-      return;
-    }
-    await provider.signAndSendTransaction();
-  };
-
-  const uiConsole = (...args: unknown[]): void => {
-    const el = document.querySelector("#console>p");
-    if (el) {
-      el.innerHTML = JSON.stringify(args || {}, null, 2);
-    }
-  };
-
-  const contextProvider = {
-    web3Auth,
-    chain,
-    provider,
-    user,
-    isLoading,
-    login,
-    logout,
-    getUserInfo,
-    getAccounts,
-    getBalance,
-    signMessage,
-    signTransaction,
-    signAndSendTransaction,
-  };
-  return <Web3AuthContext.Provider value={contextProvider}>{children}</Web3AuthContext.Provider>;
-};
